requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
filelock==3.13.1
fsspec==2023.10.0
ibm-cloud-sdk-core==3.18.0
ibm-watson==7.0.1
idna==3.5
imageio==2.33.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
Pillow==10.1.0
proglog==0.1.10
PyJWT==2.8.0
python-dateutil==2.8.2
regex==2023.10.3
requests==2.31.0
six==1.16.0
srt==3.5.3
sympy==1.12
tiktoken==0.5.1
torch==2.1.1
tqdm==4.66.1
triton==2.1.0
typing-extensions==4.8.0
urllib3==1.26.18
websocket-client==1.6.4
zipp==3.17.0
